+++ dist-test/node-cjs-ext.js
@@ -1,4 +1,5 @@
-const { render } = require('../dist/micromustache.js')
+// This is a "normal" Node js file which tries to load the CommonJS version specifying its extension
+const { render } = require('../dist/micromustache.cjs')
 
 const result = render('Hello {{ name }}!', { name: 'Alex' })
 if (result !== 'Hello Alex!') {