+++ dist-test/node-cjs-no-ext.js
@@ -1,4 +1,5 @@
-const { render } = require('../dist/micromustache')
+// This is a Node CommonJS file which tries to use the named CommonJS export as specified in the package.json
+const { render } = require('..')
 
 const result = render('Hello {{ name }}!', { name: 'Alex' })
 if (result !== 'Hello Alex!') {