+++ dist-test/node-esm-use-cjs.mjs
@@ -1,4 +1,5 @@
-import micromustache from '../dist/micromustache.js'
+// This is a node ESM module which tries to access the CommonJS build for some reason!
+import micromustache from '../dist/micromustache.cjs'
 
 const { render } = micromustache
 const result = render('Hello {{ name }}!', { name: 'Alex' })