+++ dist-test/node-esm.mjs
@@ -1,6 +1,6 @@
-import micromustache from '../dist/micromustache'
+// This is a Node ESM file which tries to import the ESM version of the library
+import { render } from '../dist/micromustache'
 
-const { render } = micromustache
 const result = render('Hello {{ name }}!', { name: 'Alex' })
 if (result !== 'Hello Alex!') {
     throw new Error(`The result does not match ${result}`)