+++ rollup.config.js
@@ -2,57 +2,59 @@ import { terser } from 'rollup-plugin-terser'
 import typescript from '@rollup/plugin-typescript'
 import { name, version } from './package.json'
 
+if (typeof name !== 'string' || typeof version !== 'string') {
+  console.log('Invalid name or version')
+  process.exit()
+}
+
 /*
 typescript and tslib are peer dependencies of @rollup/plugin-typescript
 */
-const banner = `/* ${name} v${version} */`
-const sourcemap = true
+const commonConfig = {
+  banner: `/* ${name} v${version} */`,
+  sourcemap: true,
+}
 
 export default [
   {
     input: 'src/index.ts',
     plugins: [
       typescript({
-        // These options are what you would pass to compilerOptions
-        tsconfig: './tsconfig-build.json',
-      }),
-      terser({
-        include: /\.min\.$/,
-        safari10: true,
+        rootDir: 'src',
+        exclude: [
+          'src/**/*.spec.*'
+        ],
       }),
     ],
     output: [
       {
-        file: `dist/${name}.js`,
         format: 'cjs',
-        banner,
-        sourcemap,
+        file: `dist/${name}.cjs`,
+        ...commonConfig,
       },
       {
+        format: 'umd',
         file: `dist/${name}.umd.js`,
         name,
-        format: 'umd',
-        banner,
-        sourcemap,
+        ...commonConfig,
       },
       {
+        format: 'umd',
         file: `dist/${name}.umd.min.js`,
         name,
-        format: 'umd',
-        banner,
-        sourcemap,
+        plugins: [terser()],
+        ...commonConfig,
       },
       {
+        format: 'esm',
         file: `dist/${name}.mjs`,
-        format: 'es',
-        banner,
-        sourcemap,
+        ...commonConfig,
       },
       {
+        format: 'esm',
         file: `dist/${name}.min.mjs`,
-        format: 'es',
-        banner,
-        sourcemap,
+        plugins: [terser()],
+        ...commonConfig,
       },
     ],
   },